requirements.txt
Flask==2.0.1
ultralytics==8.0.229
requests==2.26.0
Pillow==8.3.2
